[Rename] __tests__ unit 폴더 구조 변경
